Adds breaking test for n_jobs passed "auto"
